requirements.txt
colorama==0.4.6
psutil==7.0.0
setproctitle==1.2.2
setuptools==80.9.0
wheel==0.45.1
pip==25.1.1
wcwidth==0.2.13
triton==3.2.0
sentencepiece==0.2.0
pytz==2025.2
py-cpuinfo==9.0.0
pure_eval==0.2.3
ptyprocess==0.7.0
nvidia-cusparselt-cu12==0.6.2
nanobind==2.6.1
mpmath==1.3.0
fastrlock==0.8.3
blake3==1.0.4
zipp==3.21.0
zipp==3.23.0
xxhash==3.5.0
widgetsnbextension==4.0.13
websockets==15.0.1
uvloop==0.21.0
urllib3==2.3.0
unsloth==2025.5.7
tzdata==2025.2
typing_extensions==4.13.1
typing_extensions==4.14.0
traitlets==5.14.3
tqdm==4.67.1
threadpoolctl==3.6.0
tabulate==0.9.0
sympy==1.13.1
sniffio==1.3.1
smmap==5.0.2
six==1.17.0
shtab==1.7.1
shellingham==1.5.4
setproctitle==1.3.5
safetensors==0.5.3
rpds-py==0.24.0
regex==2024.11.6
pyzmq==26.4.0
pyzmq==27.0.0
PyYAML==6.0.2
python-multipart==0.0.20
python-json-logger==3.3.0
python-dotenv==1.1.0
pyparsing==3.2.1
Pygments==2.19.1
pycountry==24.6.1
pybind11==2.13.6
pyarrow==19.0.1
psutil==7.0.0
protobuf==3.20.3
propcache==0.3.1
prompt_toolkit==3.0.51
prometheus_client==0.21.1
pluggy==1.5.0
platformdirs==4.3.8
pillow==11.1.0
pexpect==4.9.0
partial-json-parser==0.2.1.1.post5
parso==0.8.4
packaging==24.2
packaging==25.0
nvidia-nvtx-cu12==12.4.127
nvidia-nvjitlink-cu12==12.4.127
nvidia-nccl-cu12==2.21.5
nvidia-curand-cu12==10.3.5.147
nvidia-cufft-cu12==11.2.1.3
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cublas-cu12==12.4.5.8
numpy==1.26.4
ninja==1.11.1.4
networkx==3.4.2
nest_asyncio==1.6.0
multidict==6.3.2
msgspec==0.19.0
msgpack==1.1.0
mdurl==0.1.2
MarkupSafe==3.0.2
llvmlite==0.43.0
llguidance==0.7.13
lark==1.2.2
kiwisolver==1.4.8
jupyterlab_widgets==3.0.13
joblib==1.4.2
jiter==0.9.0
interegular==0.3.3
iniconfig==2.0.0
idna==3.10
httptools==0.6.4
hf_transfer==0.1.9
h11==0.14.0
fsspec==2024.12.0
frozenlist==1.5.0
fonttools==4.56.0
filelock==3.18.0
executing==2.2.0
emoji==2.14.1
einops==0.8.1
docstring_parser==0.16
dnspython==2.7.0
distro==1.9.0
diskcache==5.6.3
dill==0.3.8
decorator==5.2.1
cycler==0.12.1
conda-pack==0.8.1
cloudpickle==3.1.1
click==8.1.8
charset-normalizer==3.4.1
certifi==2025.1.31
cachetools==5.5.2
attrs==25.3.0
asttokens==3.0.0
astor==0.8.1
annotated-types==0.7.0
airportsdata==20250224
aiohappyeyeballs==2.6.1
absl-py==2.2.1
yarl==1.19.0
uvicorn==0.34.0
typing-inspection==0.4.0
typeguard==4.4.2
stack_data==0.6.3
sentry-sdk==2.23.1
scipy==1.15.2
requests==2.32.3
referencing==0.36.2
python-dateutil==2.9.0.post0
pytest==8.3.5
pydantic_core==2.33.1
opencv-python-headless==4.11.0.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-cudnn-cu12==9.1.0.70
numba==0.60.0
nltk==3.9.1
multiprocess==0.70.16
matplotlib-inline==0.1.7
markdown-it-py==3.0.0
Jinja2==3.1.6
jedi==0.19.2
ipython_pygments_lexers==1.1.1
importlib_metadata==8.6.1
importlib_metadata==8.7.0
imojify==0.1.1
httpcore==1.0.7
gitdb==4.0.12
gguf==0.10.0
email_validator==2.2.0
docker-pycreds==0.4.0
depyf==0.18.0
cupy-cuda12x==13.4.1
contourpy==1.3.1
comm==0.2.2
anyio==4.9.0
aiosignal==1.3.2
watchfiles==1.0.4
tiktoken==0.9.0
starlette==0.46.1
scikit-learn==1.6.1
rouge_score==0.1.2
rich==14.0.0
pydantic==2.11.2
pandas==2.2.3
nvidia-cusolver-cu12==11.6.1.9
matplotlib==3.10.1
jsonschema-specifications==2024.10.1
ipython==9.3.0
smolagents==1.18.0
httpx==0.28.1
GitPython==3.1.44
aiohttp==3.11.16
wandb==0.19.9
UpSetPlot==0.9.0
tyro==0.9.18
typer==0.15.2
torch==2.6.0
tokenizers==0.21.1
seaborn==0.13.2
rich-toolkit==0.14.1
prometheus-fastapi-instrumentator==7.1.0
openai==1.70.0
lm-format-enforcer==0.10.11
jsonschema==4.23.0
iterative-stratification==0.1.9
ipywidgets==8.1.5
fastapi==0.115.12
diffusers==0.32.2
xformers==0.0.29.post2
transformers==4.51.3
torchvision==0.21.0
torchaudio==2.6.0
ray==2.43.0
outlines_core==0.1.26
mistral_common==1.5.4
fastapi-cli==0.0.7
datasets==3.5.0
cut-cross-entropy==25.1.1
bitsandbytes==0.46.0
accelerate==1.6.0
xgrammar==0.1.16
trl==0.15.2
peft==0.15.1
outlines==0.1.11
compressed-tensors==0.9.2
unsloth_zoo==2025.5.8
vllm==0.8.2
debugpy==1.8.14
pickleshare==0.7.5
tornado==6.5.1
exceptiongroup==1.3.0
jupyter_core==5.8.1
jupyter_client==8.6.3
ipykernel==6.29.5
hf-xet==1.1.5
huggingface-hub==0.33.0
cmap==0.6.2
autocommand==2.2.2
backports.tarfile==1.2.0
importlib_metadata==8.0.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
more-itertools==10.3.0
packaging==24.2
platformdirs==4.2.2
tomli==2.0.1
typeguard==4.3.0
typing_extensions==4.12.2
wheel==0.45.1
zipp==3.19.2
